more flash match info. not fully tested
